+++ src/event_handler/overlays.rs
@@ -5,12 +5,12 @@ pub(super) fn in_overlay_mode(app: &App) -> bool {
 }
 
 pub(super) fn select_searched_service(app: &mut App) {
-    if let Some(index) = app.services.iter().position(|service| {
-        service
-            .name
-            .to_lowercase()
-            .starts_with(&app.search_query.to_lowercase())
-    }) {
+    let query = app.search_query.to_lowercase();
+    if let Some(index) = app
+        .services
+        .iter()
+        .position(|service| service.name.to_lowercase().contains(&query))
+    {
         app.state.select(Some(index));
     }
 }